Fixed numpy bug (?) and implemented Config dataclass
The Config dataclass can be used to store and retrieve configs.
It is a better approach than plain json since we get autocompletion
of the attributes on IDEs (freeing us from dummy bugs) and proper
documentation.

The numpy bug is resolved but there's no real explanation. By
squeezing and then re-adding the last dummy one dimension it
works like a charm.

Co-authored-by: beabevi <29659657+beabevi@users.noreply.github.com>