Implemented encoder-decoder model in seq2seq_keras
In this case we implemented the model of the paper i.e. the
one feeding to the decoder the past history of the target serie.
Currently the model works only on SML2010 dataset while producing bad
performances in the NASDAQ100 one.

EarlyStopping from keras is used in order to restore the best
weights (the ones producing the best validation loss) are restored
at the end of the training

The portion of code in Prediction Model section is currently unused
although it can/will be used for multi-step ahead predictions

Co-authored-by: beabevi <29659657+beabevi@users.noreply.github.com>
Co-authored-by: spallas <12670376+spallas@users.noreply.github.com>